requirements.txt
PyYAML~=6.0.1
pydantic~=2.4.1
scinum~=1.4.3
requests~=2.31.0
